Renamed lib to core in clone command
