+++ packages/contentstack-clone/src/core/util/dummyConfig.json
