+++ packages/contentstack-clone/test/commands/stack-clone.test.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
-import { CloneHandler } from '../../src/lib/util/clone-handler';
+import { CloneHandler } from '../../src/core/util/clone-handler';
 import { CloneConfig } from '../../src/types/clone-config';
 import inquirer from 'inquirer';
 // eslint-disable-next-line @typescript-eslint/no-var-requires
